Custom Error Fix

+++ ReactAndroid/src/main/java/com/facebook/react/animated/ModulusAnimatedNode.java
@@ -1,8 +1,10 @@
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
  *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
  */
 
 package com.facebook.react.animated;
@@ -15,14 +17,14 @@ import com.facebook.react.bridge.ReadableMap;
 
   private final NativeAnimatedNodesManager mNativeAnimatedNodesManager;
   private final int mInputNode;
-  private final double mModulus;
+  private final int mModulus;
 
   public ModulusAnimatedNode(
       ReadableMap config,
       NativeAnimatedNodesManager nativeAnimatedNodesManager) {
     mNativeAnimatedNodesManager = nativeAnimatedNodesManager;
     mInputNode = config.getInt("input");
-    mModulus = config.getDouble("modulus");
+    mModulus = config.getInt("modulus");
   }
 
   @Override
